requirements.txt
streamlit
fastapi
uvicorn
python-dotenv
openai-whisper
mistralai
edge-tts
playwright
beautifulsoup4
fpdf
pdfplumber
python-multipart
python-jose[cryptography]
requests
chardet
